optimization and bug fixes

+++ src/edu/upenn/cit594/processor/AverageTotalLivableArea.java
@@ -4,6 +4,11 @@ import java.util.List;
 
 import edu.upenn.cit594.data.Property;
 
+/**
+ * 
+ * @author zhongliu
+ *
+ */
 public class AverageTotalLivableArea implements AverageInZipCode{
 	@Override
 	public double averageInZipCode(String zipCode, List<Property> propertyList) {
@@ -11,15 +16,16 @@ public class AverageTotalLivableArea implements AverageInZipCode{
 		double totalAreaValue = 0;
 		for (int i = 0; i < propertyList.size(); i++ ) {
 			String propertyZipCode = propertyList.get(i).getZipCode();
-			Integer propertyLivableArea = propertyList.get(i).getTotalLiverableArea();
+			Double propertyLivableArea = propertyList.get(i).getTotalLiverableArea();
 			if (propertyLivableArea != null && propertyZipCode != null) {
 				if (zipCode.equals(propertyZipCode)) {
 					totalAreaValue += propertyLivableArea;
 					propertyCounter++;
 				}	
 			}
-			
 		}
+//		System.out.println("totalAreaValue="+totalAreaValue);
+//		System.out.println("propertyCounter="+propertyCounter);
 		if (propertyCounter == 0) {
 			return 0;
 		}
